- Changed energy battery and health UI art + Added dynamic health UI change system + Added ammo system - Disabled Enemy Health Bar scripts on enemy prefabs - Changed Wispling damage to 1 - Changed max health and current health variables from floats to int type across scripts

+++ RogueLove/Assets/Scripts/EnemyBulletScript.cs
@@ -23,7 +23,7 @@ public class EnemyBulletScript : MonoBehaviour
 
     [SerializeField] private float force;
 
-    [SerializeField] private float damage = 2;
+    [SerializeField] private int damage = 2;
 
     [SerializeField] private bool reflected = false;
